+++ app/views/candidate_interface/course_choices/review_and_submit/show.html.erb
